fix function - condition

+++ 0x1A-hash_tables/4-hash_table_get.c
@@ -8,7 +8,7 @@
 char *hash_table_get(const hash_table_t *ht, const char *key)
 {
 	hash_node_t *hn = NULL;
-	int index;
+	unsigned long int index;
 
 	if (ht == NULL && key == NULL)
 		return (NULL);
